sfsq: carry matched count as u64, narrow to usize at the wire
The match count is a bitmap cardinality (u64) the whole way up — sfst's
matched_count returns u64, LogsShard sums in u64 — but LogsData narrowed
it to usize via an early `as usize` inside the engine. That made matched
the only count cast before the wire boundary: facet (u32) and histogram
bucket (u64) counts are carried sfst-native in LogsData and cast to usize
at the adapter (adapter.rs:226, 284-285).

Make matched follow the buckets: LogsData.matched is now u64 (matching
the Bucket::counts it already carries), the engine drops its cast, and the
adapter does `data.matched as usize` next to the existing bucket casts.
The u64->usize narrowing now lives at the single UI boundary every count
shares instead of hidden mid-engine. The wire Items.matched stays usize;
no behavior change on the 64-bit target.

+++ src/crates/otel-ledger/src/ledger/rpc/adapter.rs
@@ -173,7 +173,7 @@ pub fn to_result(data: LogsData, max_to_return: usize) -> LogsResult {
     let available_histograms = available_histograms_from_fields(&data.available_fields);
     let facetable = data.facetable();
     let (columns, rows) = build_table(&data.rows, &data.columns, &facetable);
-    let matched = data.matched;
+    let matched = data.matched as usize;
 
     LogsResult {
         progress: 100,
